Fix reload bugs
Fixed bug in Gun.cs that stopped reload from being called. This was due to the original magazine size being used without adjusting for modifiers.

Fixed bug in GunReloadComponent.cs that caused reload times to be twice as long. This was caused due to multiplication being used instead of division.