Fix spread over time crosshair size
Crosshair was previously only resizing to a radius of 0 after the first shot and reset, this was due projectileSpreadIncrementValue not being set when the spread over time mod is initialised. Values when calculating the crosshair size were also incorrect and used obsolete code.

Updated Gun.cs to initialize mods in Awake.

ProjectileSpreadOverTimeMod.cs now sets the value to 0f when initialised.

Removed spreadRadiusModifier as this was legacy code from before projectileSpreadIncrementValue was created.

Fixed SetCrosshairSize method in gun.cs as previously calculating on obsolete code.